- Added RB_ShowTime() class - Changed ScreenWidget() class - In RollingBallGameMode() class changed Tick(), BeginPlay(), GameOver() and EndGame() functions

+++ Source/RollingBall/Public/RB_SaveGame.h
@@ -18,7 +18,7 @@ class ROLLINGBALL_API URB_SaveGame : public USaveGame
 public:
 
 	UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Time")
-	float GameTime;
+	int32 GameTime;
 
 	// ******** FUNCTIONS ******** //
